Break out inventory fetch and push
it turns out you can make post requests you just have to add a special header..
